Change result type to `f64`

+++ src/mathrs.rs
@@ -6,7 +6,7 @@ pub use lexer::Lexer;
 pub use parser::Parser;
 pub use interpreter::interpret;
 
-pub fn eval(text: &mut str) -> Option<i32> {
+pub fn eval(text: &mut str) -> Option<f64> {
     let mut p = Parser::new(text);
     let ast = p.parse_expr()?;
     Some(interpret(ast))
